changed newThought shortcut to genric enter shortcut
updated the enter test

migrated newThought shorcuts exec logic to a new action creator
