+++ src/action-creators/index.js
@@ -14,6 +14,7 @@ export { default as loadRemoteState } from './loadRemoteState'
 export { default as loadResource } from './loadResource'
 export { default as login } from './login'
 export { default as logout } from './logout'
+export { default as newThought } from './newThought'
 export { default as preloadSources } from './preloadSources'
 export { default as setEditingValue } from './setEditingValue'
 export { default as setInvalidState } from './setInvalidState'
