+++ src/shortcuts/index.js
@@ -18,7 +18,7 @@ export { default as moveThoughtDown } from './moveThoughtDown'
 export { default as moveThoughtUp } from './moveThoughtUp'
 export { default as newSubthought, newSubthoughtAliases } from './newSubthought'
 export { default as newSubthoughtTop } from './newSubthoughtTop'
-export { default as newThought, newThoughtAliases } from './newThought'
+export { default as enter, enterAliases } from './enter'
 export { default as newThoughtAbove } from './newThoughtAbove'
 export { default as newUncle } from './newUncle'
 export { default as note } from './note'
